pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp_examples"
version = "0.1.0"
authors = [{ name = "yu-iskw" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11.0"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
description = "A dbt artifacts parser in python"
dependencies = [
  "duckduckgo-search>=7.4.2",
  "google-genai>=1.2.0",
  "httpx>=0.28.1",
  "langgraph>=0.2.73",
  "loguru>=0.7.3",
  "mcp[cli]>=1.2.1",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
]

[project.optional-dependencies]
dev = ["black>=25.1.0", "isort>=6.0.0", "pytest>=8.3.4", "ruff>=0.9.6"]
